--- packages/create-package/template/config/postcss.components/postcss.config.js
@@ -1,1 +1,0 @@
-module.exports = require("@ui5/webcomponents-tools/components-package/postcss.components.js"); // eslint-disable-line
