--- packages/create-package/template/config/postcss.themes/postcss.config.js
@@ -1,1 +1,0 @@
-module.exports = require("@ui5/webcomponents-tools/components-package/postcss.themes.js"); // eslint-disable-line
