--- packages/main/config/postcss.components/postcss.config.cjs
@@ -1,1 +1,0 @@
-module.exports = require("@ui5/webcomponents-tools/components-package/postcss.components.js");
