--- packages/main/config/postcss.themes/postcss.config.cjs
@@ -1,1 +1,0 @@
-module.exports = require("@ui5/webcomponents-tools/components-package/postcss.themes.js");
