+++ packages/theming/package.json
@@ -35,7 +35,6 @@
   },
   "devDependencies": {
     "@ui5/webcomponents-tools": "1.21.0-rc.2",
-    "cssnano": "^6.0.1",
     "globby": "^13.1.1",
     "json-beautify": "^1.1.1",
     "nps": "^5.10.0",
