+++ packages/tools/package.json
@@ -36,7 +36,6 @@
     "command-line-args": "^5.1.1",
     "concurrently": "^6.0.0",
     "cross-env": "^7.0.3",
-    "cssnano": "^6.0.1",
     "escodegen": "^2.0.0",
     "eslint": "^7.22.0",
     "eslint-config-airbnb-base": "^14.2.1",
@@ -54,7 +53,6 @@
     "nps": "^5.10.0",
     "postcss": "^8.4.5",
     "postcss-cli": "^9.1.0",
-    "postcss-import": "^14.0.2",
     "postcss-selector-parser": "^6.0.10",
     "prompts": "^2.4.2",
     "properties-reader": "^2.2.0",
@@ -63,8 +61,7 @@
     "rimraf": "^3.0.2",
     "slash": "3.0.0",
     "vite": "^4.4.9",
-    "wdio-chromedriver-service": "^7.3.2",
-    "zx": "^7.2.3"
+    "wdio-chromedriver-service": "^7.3.2"
   },
   "peerDependencies": {
     "chromedriver": "*",
